+++ apps/api/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @embedly/api
 
+## 0.1.3
+
+### Patch Changes
+
+- fixed tiktok embeds and added new instagram share link handling
+- Updated dependencies
+  - @embedly/platforms@0.1.3
+  - @embedly/parser@0.1.3
+  - @embedly/builder@0.1.3
+  - @embedly/logging@0.1.3
+  - @embedly/types@0.1.3
+
 ## 0.1.2
 
 ### Patch Changes