+++ packages/config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @embedly/config
 
+## 0.1.3
+
+### Patch Changes
+
+- fixed tiktok embeds and added new instagram share link handling
+
 ## 0.1.2
 
 ### Patch Changes