+++ packages/parser/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @embedly/parser
 
+## 0.1.3
+
+### Patch Changes
+
+- fixed tiktok embeds and added new instagram share link handling
+- Updated dependencies
+  - @embedly/types@0.1.3
+
 ## 0.1.2
 
 ### Patch Changes